Remove Golioth from rasprover and add it back to builds
Golioth-free rasprover now builds stand-alone; Golioth can be
re-enabled later once the manifest modules are available.

rasprover source:
- Remove golioth/client, golioth/fw_update, golioth/stream, golioth/settings
  includes and all dependent logic from main.c, app_sensors.c/h, app_settings.c/h
- app_sensors: read and log INA219 data locally instead of streaming to Golioth
- app_settings: keep loop-delay and settings subsystem init; drop PSK credential
  handling and Golioth settings registration
- app_network: update net_mgmt event handler signatures uint32_t -> uint64_t
  (Zephyr API change); fix dhcpv4.h / net_if.h include order
- prj.conf: drop all CONFIG_GOLIOTH_* and OTA/firmware-upgrade config
- sysbuild.conf: remove SB_CONFIG_BOOTLOADER_MCUBOOT (no longer needed)
- overlay: update SSD1306 compatible solomon,ssd1306fb -> solomon,ssd1306
  (old binding removed from Zephyr main)

CI: add rasprover / ros_driver/esp32/procpu to build matrix
poe.toml: add build-rasprover task

+++ applications/rasprover/src/app_network.c
@@ -3,8 +3,8 @@ LOG_MODULE_REGISTER(app_network, LOG_LEVEL_INF);
 
 #include <zephyr/device.h>
 #include <zephyr/kernel.h>
-#include <zephyr/net/dhcpv4.h>
 #include <zephyr/net/net_if.h>
+#include <zephyr/net/dhcpv4.h>
 #include <zephyr/net/net_mgmt.h>
 #include <zephyr/net/wifi_mgmt.h>
 #include <zephyr/pm/device.h>
@@ -408,7 +408,7 @@ static inline atomic_val_t atomic_update(atomic_t *target, atomic_val_t value, a
 	return flags;
 }
 
-static void ipv4_changed(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+static void ipv4_changed(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 			 struct net_if *iface)
 {
 	struct wifi_manager_data *wifi_mgmt =
@@ -433,7 +433,7 @@ static inline int wifi_connect_error(const struct wifi_status *status)
 	return status->status;
 }
 
-static void wifi_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+static void wifi_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 				    struct net_if *iface)
 {
 	struct wifi_manager_data *wifi_mgmt =
@@ -512,7 +512,7 @@ struct wait_data {
 	struct net_mgmt_event_callback cb;
 };
 
-static void event_cb_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+static void event_cb_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 			     struct net_if *iface)
 {
 	struct wait_data *wait = CONTAINER_OF(cb, struct wait_data, cb);
@@ -522,7 +522,7 @@ static void event_cb_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_e
 	}
 }
 
-static void wait_for_net_event(struct net_if *iface, uint32_t event)
+static void wait_for_net_event(struct net_if *iface, uint64_t event)
 {
 	struct wait_data wait;
